requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.0.0
playwright==1.42.0
supabase==2.0.3
openpyxl==3.1.2
pandas==2.0.3
numpy>=1.26.0
python-dateutil>=2.8.2
pytz>=2020.1
tzdata>=2022.7 